docs: :art: change pseudo code blocks including commit messages from bash to default
By changing these blocks to "default", all text in the code blocks is now the same colour. E.g., the opening paranthesis in commit messages including scope were shown in red in bash code blocks. By changing it to "default", the parenthesis is now the same colour as the rest of the text in the block.