docs: :memo: update iteration.qmd (#227)
## Description

This PR updates the iteration docs based on how we currently work.

Note that there might be more things to update than this document, but
these are some of the things that come up when we look at the doc during
planning and debrief/retrospective meetings.

<!-- Please delete as appropriate: -->
This PR needs a quick review.

## Checklist

- [X] Ran spell-check
- [X] Formatted Markdown
- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: Kris Beicher <112945740+K-Beicher@users.noreply.github.com>